E2E: move 8.0 workaround to mutation steps
