+++ test/e2e/test/elasticsearch/steps_creation.go
@@ -7,10 +7,6 @@ package elasticsearch
 import (
 	"context"
 	"fmt"
-	"testing"
-	"time"
-
-	"github.com/elastic/cloud-on-k8s/pkg/controller/common/version"
 
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp" // auth on gke
 
@@ -41,16 +37,6 @@ func (b Builder) CreationTestSteps(k *test.K8sClient) test.StepList {
 					// TODO this is incomplete
 					return nil
 				}),
-			},
-			test.Step{
-				Name: "Give Elasticsearch some time to allocate internal indices",
-				Test: func(_ *testing.T) {
-					// TODO remove this step once https://github.com/elastic/cloud-on-k8s/issues/5040 does not apply anymore
-					time.Sleep(30 * time.Second)
-				},
-				Skip: func() bool {
-					return version.MustParse(b.Elasticsearch.Spec.Version).LT(version.MinFor(7, 16, 0))
-				},
 			},
 		})
 }
